Split bearing into two files for better testing
